backend/requirements.txt
flask
flask-socketio
python-socketio
python-engineio
requests
python-telegram-bot==20.3
eventlet
